Made code more oop, added relevant classes as required
