+++ text-ui-test/input.txt
@@ -6,4 +6,4 @@ mark 2
 list
 unmark 2
 list
-bye
+exit
